Add: IBC types in Registry
We missed ibc types in our Protobuf Registry because of which,
governance tab was not showing any proposals (old or new) and was
failing to load any information. The main issue was with
`/ibc.core.client.v1.ClientUpdateProposal` type and is now being
registered with other common IBC data types.

Signed-off-by: jay-dee7 <jasdeepsingh.uppal@gmail.com>