Fix: vesting calculation integer overflow
We were using a lot of parseInt methods in estimatedVesting method and
it was leading to integer overflow for large numbers. We've switched
from parsing ints of `Coin.amount` to just type casting the amount
(which is a string) to Number in TS, which seems to fix the issue as of
now.

Signed-off-by: jay-dee7 <jasdeepsingh.uppal@gmail.com>